M3 v1.0 (alpha) and documentation
The first 'fit for purpose' versiono of the M3 operating model is now
available.

This includes an updated software manual and software design document
(/Manuals and design documents/).

Progress report 5 is currently being drafted.

The latest version of the M3 .pdf report is available but yet to include
the statistical properties of fits with relative abundance indices to be
used by prospective MPs (e.g. correlation, autocorrelation of the
Balearic larval index with Med SSB)